// src/app/core/models/si/db/mapping-destination.mode.ts
import { DestinationAttributeDetail } from "./destination-attribute.model";

export type MappingDestination = {
  id: number;
  attribute_type: string;
  display_name: string;
  value: string;
  destination_id: string;
  active: boolean;
  created_at: Date;
  updated_at: Date;
  workspace: number;
  detail: DestinationAttributeDetail;
};